put all of simplegrants into one file (netlist.py). Same for tests.
